tests: replace test_text_plain_and_other
this created a funny message with an "application/octet-stream" part
that was no attachment, and tested if this part was ignored.

It should (and now is) not ignored and instead raises an error
mentioning that we tried and failed to find a mailcap entry.
A new test is added here to verify this behaviour.